move ReviewChanges to separate file
